Fix the inputValue in the docs.
When using the storybook UI we couldn't change the inputs values directly
by typing on them.
The problem is an issue with the knob addon. If we add a value to the input
with knob, storybook won't allow the user changing the value directly,
only if the user uses the knob tab.